% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp_brands.r
\docType{data}
\name{airam_lamps}
\alias{airam_lamps}
\title{Spectral data for lamps supplied by Airam}
\format{A vector of character strings.}
\usage{
airam_lamps
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for lamps from Airam Electric Oy Ab. 
Data are normalized to one at the wavelength of maximum emission.
}
\examples{
osram_lamps

}
\references{
\url{https://www.airam.fi/en/}
}
\seealso{
\code{\link{lamps.mspct}}

Other manufacturers: \code{\link{aputure_lamps}},
  \code{\link{generic_lamps}}, \code{\link{osram_lamps}},
  \code{\link{philips_lamps}}, \code{\link{qpanel_lamps}},
  \code{\link{sylvania_lamps}},
  \code{\link{toshiba_lamps}}, \code{\link{valoya_lamps}}
}
\keyword{datasets}
