% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp-brands.r
\docType{data}
\name{aputure_lamps}
\alias{aputure_lamps}
\title{Spectral data for lamps supplied by Aputure}
\format{A vector of character strings.}
\usage{
aputure_lamps
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for photography and video lamps from Aputure. 
Data are normalized to one at the wavelength of maximum emission.
}
\examples{
aputure_lamps

}
\references{
\url{https://www.aputure.com/}
}
\seealso{
\code{\link{lamps.mspct}}

Other manufacturers: \code{\link{airam_lamps}},
  \code{\link{fluence_lamps}}, \code{\link{generic_lamps}},
  \code{\link{godox_lamps}}, \code{\link{osram_lamps}},
  \code{\link{philips_lamps}}, \code{\link{qpanel_lamps}},
  \code{\link{sunwayfoto_lamps}},
  \code{\link{sylvania_lamps}},
  \code{\link{toshiba_lamps}}, \code{\link{valoya_lamps}}
}
\concept{manufacturers}
\keyword{datasets}
