% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp_brands.r
\docType{data}
\name{generic_lamps}
\alias{generic_lamps}
\title{Spectral data for shop-branded lamps}
\format{A vector of character strings.}
\usage{
generic_lamps
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for lamps from Toshiba. 
Data are normalized to one at the wavelength of maximum emission.
}
\examples{
generic_lamps

}
\seealso{
\code{\link{lamps.mspct}}

Other manufacturers: \code{\link{airam_lamps}},
  \code{\link{aputure_lamps}}, \code{\link{osram_lamps}},
  \code{\link{philips_lamps}}, \code{\link{qpanel_lamps}},
  \code{\link{sylvania_lamps}},
  \code{\link{toshiba_lamps}}, \code{\link{valoya_lamps}}
}
\keyword{datasets}
