% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macam.r
\docType{data}
\name{macam_lamps}
\alias{macam_lamps}
\title{Spectra acquired with Macam SR-9010-PC}
\format{A vector of character strings.}
\usage{
macam_lamps
}
\description{
Lamp emission spectra measured with a Macam
  SR-9010-PC spectroradiometer.
}
\details{
\code{macam_lamps} is a character vector with indexes to members of 
  \code{\link{lamps.mspct}}.
}
\note{
Instrument used: Macam SR-9010-PC scanning double monochromator
  spectroradiometer with a cosine corrected input optics. Recently
  calibrated. The lamps were probably driven by a high frequency electronic
  ballast. Measured between 1996 to 1998.
}
\examples{
macam_lamps

}
\references{
Aphalo, P J, R Tegelberg, and R Julkunen-Tiitto. 1999. The
  Modulated UV-B Irradiation System at the University of Joensuu. Biotronics
  28: 109-120.
  \url{http://ci.nii.ac.jp/naid/110006175827/en}.
}
\keyword{datasets}
