% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uses.r
\docType{data}
\name{photography_lamps}
\alias{photography_lamps}
\alias{uses}
\alias{plant_grow_lamps}
\title{Spectral data for Lamps designed for specific uses}
\format{A vector of character strings.}
\usage{
photography_lamps

plant_grow_lamps
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for the lamps grouped by special use.
}
\note{
In the case of LED-based lamps we include here only ready built commercially 
available lamps.
In some cases assembled from multiple descrete LEDs, possible of mixed types and 
spectral output. For emission spectra for LEDs available as electronic components
please see \code{\link[photobiologyLEDs]{photobiologyLEDs-package}}.
}
\examples{
plant_grow_lamps
photography_lamps

# select lamps for photography
lamps.mspct[photography_lamps]

}
\seealso{
\code{\link{lamps.mspct}}
}
\keyword{datasets}
