% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp-brands.r
\docType{data}
\name{qpanel_lamps}
\alias{qpanel_lamps}
\title{Spectral data for lamps supplied by Q-LAB}
\format{A vector of character strings.}
\usage{
qpanel_lamps
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for lamps branded Q-Panel and supplied by Q-LAB. 
Data are normalized to one at the wavelength of maximum emission.
}
\examples{
qpanel_lamps

}
\references{
\url{http://www.q-lab.com/}
}
\seealso{
\code{\link{lamps.mspct}}

Other manufacturers: \code{\link{airam_lamps}},
  \code{\link{aputure_lamps}}, \code{\link{fluence_lamps}},
  \code{\link{generic_lamps}}, \code{\link{godox_lamps}},
  \code{\link{osram_lamps}}, \code{\link{philips_lamps}},
  \code{\link{sunwayfoto_lamps}},
  \code{\link{sylvania_lamps}},
  \code{\link{toshiba_lamps}}, \code{\link{valoya_lamps}}
}
\concept{manufacturers}
\keyword{datasets}
