% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp-brands.r
\docType{data}
\name{lamp_brands}
\alias{lamp_brands}
\alias{osram_lamps}
\alias{philips_lamps}
\alias{sylvania_lamps}
\alias{qpanel_lamps}
\alias{airam_lamps}
\alias{toshiba_lamps}
\alias{aputure_lamps}
\alias{sunwayfoto_lamps}
\alias{godox_lamps}
\alias{valoya_lamps}
\alias{fluence_lamps}
\alias{convoy_lamps}
\alias{jaxman_lamps}
\alias{generic_lamps}
\title{Spectral data for Lamps from different suppliers}
\format{
A vector of character strings.

An object of class \code{character} of length 8.

An object of class \code{character} of length 15.

An object of class \code{character} of length 1.

An object of class \code{character} of length 2.

An object of class \code{character} of length 4.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 3.

An object of class \code{character} of length 5.

An object of class \code{character} of length 1.

An object of class \code{character} of length 7.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.
}
\usage{
lamp_brands

osram_lamps

philips_lamps

sylvania_lamps

qpanel_lamps

airam_lamps

toshiba_lamps

aputure_lamps

sunwayfoto_lamps

godox_lamps

valoya_lamps

fluence_lamps

convoy_lamps

jaxman_lamps

generic_lamps
}
\description{
The collection of spectra \code{\link{lamps.mspct}} contains spectra for 
light bulds, luminaires and other ready-to-use lamps from various current
and former suppliers. The character
vectors described here contain the names of the spectra for lamps from each
brand to facilitate their extraction from the collection. One
additional vector, \code{lamp_brands} contains the names of the brands as used
in the names of the spectra in the collection.
}
\note{
Some of the lamps are from resellers' brands, usually obscure ones, and
  others are from well-known brands. Some of the brands have changed 
  ownership through take-overs and merges. When known the vintage is 
  indicated in the comment field of each spectrum. Multiple brands may also
  belong to the same company.
}
\section{Osram}{

Lamps branded Osram are currently sold by Ledvance, of Chinese ownership.
Osram currently makes and sells other products. When Osram (Germany) and
Sylvania (USA) merged years back both brands remained in use in different
markets and many types of lamps remained differentiated.
 .
}

\section{Philips}{

Lamps branded Philips are sold by Signify Netherlands B.V.
}

\section{Sylvania}{

Lamps branded Sylvania are currently sold by Ledvance, of Chinese ownership.
Osram currently makes and sells other products. When Osram (Germany) and
Sylvania (USA) merged years back both brands remained in use in different
markets and many types of lamps remained differentiated.
}

\section{Q-LAB}{

Q-Lab Corporation is a global provider of material durability testing 
products including UV-B and UV-A lamps used in test cabinets.
}

\section{Airam}{

Airam Electric Oy Ab is Finnish supplier and manufacturer of lamps and
luminaires.
}

\section{Toshiba}{

Lamps branded Toshiba are sold by Toshiba Lighting.
}

\section{Aputure}{

Aputure is a Chinese supplier of LED lighting and related equipment for filmmaking.
}

\section{Sunwayfoto}{

Sunwayfoto is a Chinese supplier of ancillary equipment for photography,
including battery-powered fill-in LED lights with high CRI ratings..
}

\section{Godox}{

Godox is a Chinese supplier of photography and video lighting equipment. They
sell both LED lights and Xenon flashes of a wide range to output power.
}

\section{Valoya}{

Valoya Oy (Finland) sells LED lamps and luminaires for commercial plant 
cultivation. Both for indoor cultivation and to supplement sunlight in
greenhouses.
}

\section{Fluence}{

Fluence, a division of Osram, sells LED luminaires for commercial plant 
cultivation. Both for indoor cultivation and to supplement sunlight in
greenhouses.
}

\section{Convoy}{

Is a Chinese supplier of LED flashlights/torches and their parts. Sold
through AliExpress, eBay and Banggood.
}

\section{Jaxman}{

Is a Chinese supplier of LED flashlights/torches. Formerly sold
through AliExpress.
}

\section{Generic or shop branded}{

We group under this category shop-branded or generic (no-brand) lamps that
usually have specifications with limited details. These lamps are unlikely to
remain consistent from batch to batch or have limited availability.
}

\examples{
lamp_brands
osram_lamps

}
\references{
\url{https://www.ledvance.com/}

\url{https://www.lighting.philips.com/}

\url{https://www.sylvania.com/Pages/default.aspx}

\url{https://www.q-lab.com/}

\url{https://www.airam.fi/en/}

\url{https://www.tlt.co.jp/tlt/index_e.htm}

\url{https://www.aputure.com/}

\url{http://www.sunwayfoto.com/e_index.aspx}

\url{https://www.godox.com/}

\url{https://www.valoya.com/}

\url{https://fluence-led.com/}
}
\seealso{
\code{\link{lamps.mspct}}
}
\keyword{datasets}
