% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp_brands.r
\docType{data}
\name{valoya_lamps}
\alias{valoya_lamps}
\title{Spectral data for lamps supplied by Valoya}
\format{A vector of character strings.}
\usage{
valoya_lamps
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for photography and video lamps from Valoya 
Data are normalized to one at the wavelength of maximum emission.
}
\examples{
valoya_lamps

}
\references{
\url{https://www.valoya.com/}
}
\seealso{
\code{\link{lamps.mspct}}

Other manufacturers: \code{\link{airam_lamps}},
  \code{\link{aputure_lamps}}, \code{\link{generic_lamps}},
  \code{\link{osram_lamps}}, \code{\link{philips_lamps}},
  \code{\link{qpanel_lamps}}, \code{\link{sylvania_lamps}},
  \code{\link{toshiba_lamps}}
}
\keyword{datasets}
